MIT-0 license: Removed.
Pretty sure we can just reuse the MIT license from GLFS and we are still
the copyright holders of that license, with the same dates. The MIT-0
License also just clutters up the book and can potentially make things
confusing. If this is a problem, please make a comment on this comment
or make an issue on https://github.com/glfs-book/lfs-qol/issues.