docstubgen: improve an error message
The "error: nonexistent folder1" message dates back to the
very first version of the script committed. Now that we have
proper error handling for nonexistent folders in a different
place (commit 6a3aa5) with an error message that actually
tells you which folder is nonexistent, this check is both
redundant and less descriptive.