Dolphin: 2412 -> 2503. Fixes #76.
I have also overhauled how building is done, or more accurately:
overhauled how Dolphin is downloaded and prepped up for building.
Downloading everything with Git leads to downloading much more data than
what is needed because of Git artifacts and submodules not needed in the
build.

This should save storage, although the instructions are more complicated
and may be a tad more hectic when updating Dolphin.