X11Libre: Archived. Fixes #191.
The section will just be a maintainence burden. It is a really
controversial project and is unclear if it will even be a long-lasting
and most notably a reliable project. SDL3 is new and not many packages
use it, but it's clear its a mainstay and makes much more sense to be in
SLFS than X11Libre. XLibre broke far too many things. That does not
inspire confidence. As such, this section's getting archived.