toml++: Add CMake to Recommended.
This commit also updates linux to 6.16. s4-3b1-pc7300.toml for the
kernel-config has been deleted as it is not needed anymore. #245 still
needs to be fixed as this commit does not fix it.

I forgot I made these changes before committing, but I don't think this
commit does any harm.