Add msgraph to SLFS.
This package is useful for students and remote workers who need to
interact with the Microsoft ecosystem. OneDrive support has been added
as default in Windows 11 for a long time now, so this allows users on
Linux to be able to access files from Windows machines that get
automatically backed up to OneDrive.

Users who install this will need to rebuild gvfs in BLFS without the
-Donedrive=false option. I've added some basic usage information into
the page as well to that effect!