CUDA: Reduce installed components.
The original instructions stemmed from a fundamental understanding of
the PKGBUILD Arch uses for CUDA and how CUDA works. nsight and such
aren't necessary for the functions of nvcc. Installing those components
results in about 7G of disk usage. But now the size is close to 4G. The
main goal of CUDA in SLFS is to just enable support for CUDA in
packages like FFmpeg and OBS-Studio. Install only what we need.