*: Move what we can from zipfiles to tarballs.
Fixes #436.

We want to ensure to use the least amount of dependencies if we can.
Using zipfiles unnecessarily when tarballs exists pulls in an unneeded
dependency, unzip from libarchive. We should stick with tarballs unless
there is no other distribution for a given file or selection of files,
like ROMs. Some of these packages edited here suffer from that so the
libarchive dependency can't be eliminated fully.