*: Standardize install invocation as per #276
> [!WARNING]
> This is not tested, and is thus marked as a draft until we confirm no
> regressions are introduced by these changes.

This commit standardizes all invocations of `install`[^1]. In so doing, I
noticed a few places where instructions could be condensed, so other
commands are also modified.

[^1]: As found with `find . -name '*.xml' | xargs grep -Hn 'install -'`

This commit does not consider the `install-*` XML entities, and does not
address the changes that need to be made to `/CONTRIBUTING.md`. Nor does
it consider whitespace formatting in commands. These issues will be
addressed in future commits.